requirements.txt
openai
faiss-cpu
numpy
pandas
scikit-learn
sentence-transformers